pbainit: Use localboot to boot into OS after unlock
QCT's EFI implementation is very aggressive at removing and
semi-permanently blocking booting from a device that has earlier not
been available. This makes the whole GPT partiiton swapping that happens
when MBRDone is toggled problematic to the EFI implementation.

Instead, let's do a leap of faith on this 'localboot' command.