Add initial version of CI pipeline
This change adds a configuration file for GitLab CI/CD that instructs it
to build the image using the containerized build tools.

The pipeline is far from perfect and should probably automatically
upload the resulting image as a release asset, but as of now it at least
prevents very broken builds from being merged to main.